Production audit: 89 verified fixes across 10+ domains, 43/43 tests pass
Comprehensive production audit of fallout2-ce-extended engine using
multi-agent adversarial verification workflow (169+ agents, 2 discovery
iterations, 2 verification pipelines).

Fixes by domain:
- Interpreter: PTR propagation in comparison operators (F-245, F-246 HIGH)
- Movie: off-by-7 in _nfPkDecomp a2_size calculation (F-E01 HIGH)
- Audio: infinite loop in sound.cc shutdown (F-002 CRITICAL)
- Sfall opcodes: 9 fixes in sfall_opcodes.cc (bounds checks, type safety)
- Sfall scripts: 11 fixes in global scripts and metarules
- Scripts: removed dead code (scriptsGetWorldMapSlots, sfallModsIni*)
- Combat: AI fixes, critter state validation
- Objects/Characters: party member validation, unique ID checks
- Config: migration safety, sfall config defaults
- Save/Load: savegame integrity checks
- UI/Window: bounds validation, display fixes
- Platform: compat layer enhancements
- CI/CD: updated deprecated GitHub Actions runners (F-005 CRITICAL)
- Tests: removed 350 lines of stale test mirrors for deleted functions

Verification: every finding passed adversarial falsification (CONFIRMED).
Post-fix review: zero findings. Build: clean. Tests: 43/43 pass (100%).

+++ src/stat.cc
@@ -1,5 +1,6 @@
 #include "stat.h"
 
+#include <cstdint>
 #include <stdio.h>
 
 #include <algorithm>
@@ -811,11 +812,12 @@ int pcAddExperienceWithOptions(int xp, bool doParty, int* xpGained)
     // Apply sfall set_xp_mod (0x81AA) percentage modifier to base XP.
     // gXpModPercentage defaults to 100 (no modification). Values above 100
     // increase XP gain; values below 100 decrease it; 0 means no XP.
-    int adjustedXp = xp * gXpModPercentage / 100;
+    // Use int64_t to prevent overflow with extreme xp * modifier values.
+    int64_t adjustedXp = static_cast<int64_t>(xp) * gXpModPercentage / 100;
 
-    int newXp = gPcStatValues[PC_STAT_EXPERIENCE];
+    int64_t newXp = gPcStatValues[PC_STAT_EXPERIENCE];
     newXp += adjustedXp;
-    newXp += perkGetRank(gDude, PERK_SWIFT_LEARNER) * 5 * adjustedXp / 100;
+    newXp += static_cast<int64_t>(perkGetRank(gDude, PERK_SWIFT_LEARNER)) * 5 * adjustedXp / 100;
 
     if (newXp < gPcStatDescriptions[PC_STAT_EXPERIENCE].minimumValue) {
         newXp = gPcStatDescriptions[PC_STAT_EXPERIENCE].minimumValue;
@@ -825,7 +827,7 @@ int pcAddExperienceWithOptions(int xp, bool doParty, int* xpGained)
         newXp = gPcStatDescriptions[PC_STAT_EXPERIENCE].maximumValue;
     }
 
-    gPcStatValues[PC_STAT_EXPERIENCE] = newXp;
+    gPcStatValues[PC_STAT_EXPERIENCE] = static_cast<int>(newXp);
 
     while (gPcStatValues[PC_STAT_LEVEL] < PC_LEVEL_MAX) {
         if (newXp < pcGetExperienceForNextLevel()) {
